KSK-1 : Move parseLocalDate to KoskiUtil.
